pyproject.toml
[tool.poetry]
name = "poc"
version = "0.1.0"
description = ""
authors = ["EnigmaDoor <diego.mrnj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.3"
gunicorn = "^21.2.0"
python-json-logger = "^2.0.7"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
sqlalchemy = "^2.0.29"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
